requirements.txt
vosk==0.3.45
sounddevice==0.5.5
webrtcvad==2.0.10
keyboard==0.13.5
pystray==0.19.5
pillow==12.1.1
pywin32==311
PySide6==6.10.2
pyinstaller==6.19.0